+++ src/shamo/core/fem/__init__.py
@@ -2,9 +2,19 @@
 from .group import Group
 from .field import Field
 from .tissue import Tissue
-from .sensors.sensor import Sensor
+from .sensors.abc import SensorABC
 from .sensors.point import PointSensor
+from .sensors.surface import SurfSensorABC, CircleSensor
 
 from .fem import FEM
 
-__all__ = ["FEM", "Field", "Group", "Tissue", "Sensor", "PointSensor"]
+__all__ = [
+    "FEM",
+    "Field",
+    "Group",
+    "Tissue",
+    "SensorABC",
+    "PointSensor",
+    "SurfSensorABC",
+    "CircleSensor",
+]